fix: drop target overlay showing up while reordering slides
